app/requirements.txt
django
gunicorn
djangorestframework
python-decouple
psycopg2-binary